+++ packages/frontend/src/hooks/useWebxdcMessageSent.ts
@@ -28,7 +28,7 @@ export function notifyWebxdcMessageSent(
 /**
  * A notifier that allows components to subscribe to Webxdc message sent events.
  * For cases where DC Event 'MsgsChanged' is not sufficient
- * (like on MainScreen where we want to update AppIcons immediately when a
+ * (like in AppIcons where we want to update the list immediately when a
  * Webxdc message is sent).
  */
 export function useWebxdcMessageSentNotifier() {
